Updated discovery service and config server properties
Made change in all repos (current and config server) to have default branch be 'main' instead of 'master'.
Made updates to discovery server to reach out to config service (may be unneeded here...).
Made updates to config server to look for main branch instead of master branch and to clone on startup.