Updated flow diagram and made some changes based on observations
Changed update/delete functionality to be that of owner in events, as well as in bookings.
Changed 'delete' functionality to 'cancel', replacing DELETE method with PATCH method
Updated unit tests to account for changes.
Added TODO comments in events and booking services, showing kafka publishing points and the topics to publish to.