Removed dependency on '.env' file in root directory. Setup db initialization scripts to use bash shell to allow for env-based user/pass settings.
I didn't like the idea of the user/pass settings for all the databases being hardcoded into the sql files. I basically changed the sql files into bash scripts,
allowing for environment variable injections into scripts if desired. Learned a little about the scope of .env files in docker compose files.  Although the .env
file is imported into the environment for docker compose, it is not automatically imported into the containers. As a result, a lot of boilerplate is needed to pass them
down the chain.  Oh well.  Live and learn...