Updated oath2 client hostname to be keycloak.
Added keycloak.crt to Java cacert truststore to see if it worked.  When hostname for oauth2 client provider urls set to keycloak, it worked.
This suggests that the truststore that was added to api-gateway is not getting accesses by the necessary client components used for login...

Note: in hosts file, keycloak is mapped to localhost.